widget components and proper export
