+++ frontend/core/src/components/index.ts
@@ -6,3 +6,4 @@ export * from '@/components/core/users'
 export * from './app-store'
 export * from './crm'
 export * from './lana'
+export * from './widgets'
